SecAchievements

+++ src/main/java/BSWeb/controllers/SECController.java
@@ -25,8 +25,8 @@ public class SECController {
     private SECRepository secRepository;
     @Autowired
     private AchievementsRepository achRepository;
-    @Autowired
-    private SECAchieveRepository secAchieveRepository;
+    //@Autowired
+    //private SECAchieveRepository secAchieveRepository;
 
     @GetMapping("")
     public String sec(Model model) {
@@ -46,7 +46,7 @@ public class SECController {
         }
     }
 
-    private void prepareModelByID(Model model, Long id){
+    /*private void prepareModelByID(Model model, Long id){
         Optional<SEC> sec = secRepository.findById(id);  // fullname, decription
         Iterable<SECAchieve> secAch = secAchieveRepository.findAll();
 
@@ -66,20 +66,62 @@ public class SECController {
             model.addAttribute("title", sec.get().getTitle());
         }
         model.addAttribute("ach", ach);
+    }*/
+
+    private void prepareModelByID(Model model, Long secid){
+        model.addAttribute("secid", secid);
+
+        Optional<SEC> sec = secRepository.findById(secid);  // fullname, decription
+        if (sec.isPresent()) {
+            model.addAttribute("sec", sec.get());
+            model.addAttribute("title", sec.get().getTitle());
+        }
+
+        var outAch = new ArrayList<Achievements>();
+        Iterable<Achievements> allAch = achRepository.findAll();
+
+        for (var item : allAch){
+            if (item.getSecid() == secid) {
+                outAch.add(item);
+            }
+
+        }
+
+        model.addAttribute("ach", outAch);
     }
 
     @GetMapping("/ТИОС")
     public String secAbout1(Model model) {
-        Long id = 1L;
-        prepareModelByID(model, id);
-        return "secAboutPage";
+        Long secid = 1L;
+        prepareModelByID(model, secid);
+
+        if (user.getAccess_level() == null){
+            return "secAboutPage";
+        }
+        switch ((Integer) user.getAccess_level()) {
+            case 0: // writer
+            case 1: // leader
+                return "admin/secAboutPage";
+            default:
+                return "secAboutPage";
+        }
     }
 
     @GetMapping("/БИС")
     public String secAbout2(Model model) {
-        Long id = 2L;
-        prepareModelByID(model, id);
-        return "secAboutPage";
+        Long secid = 2L;
+        prepareModelByID(model, secid);
+
+        if (user.getAccess_level() == null){
+            return "secAboutPage";
+        }
+        switch ((Integer) user.getAccess_level()) {
+            case 0: // writer
+            case 1: // leader
+                return "admin/secAboutPage";
+            default:
+                return "secAboutPage";
+        }
     }
 
     @GetMapping("/edit")
@@ -107,35 +149,43 @@ public class SECController {
     }
 
     @PostMapping("/achievements")
-    public String addAchievement(@RequestParam("sec") Long id,
-                          @RequestParam("description") String description,
+    public String addAchievement(@RequestParam("secid") Long secid,
+                          @RequestParam("title") String title,
                           Model model){
 
-        /*Optional<SEC> sec = secRepository.findById(id);  // fullname, decription
-        Iterable<SECAchieve> secAch = secAchieveRepository.findAll();
+        if (user.getAccess_level() != null){  // проверка роли
+            Achievements achievement = new Achievements(secid, title);
+            achRepository.save(achievement);
+        }
 
-        // пока что так...
-        var achIDs = new ArrayList<Long>();
+        return "redirect:/scientific-and-educational-centers";
+    }
 
-        for (var item : secAch){
-            if (item.getSec_id() == id) {
-                achIDs.add(item.getAchieve_id());
+    @PostMapping("/achievements/edit")
+    public String editAchievement(@RequestParam("id") Long id,
+                                  @RequestParam("secid") Long secid,
+                                  @RequestParam("title") String title,
+                                  Model model){
+
+        if (user.getAccess_level() != null){  // проверка роли
+            if(achRepository.existsById(id)) {
+                Achievements achievement = new Achievements(id, secid, title);
+                achRepository.save(achievement);
             }
         }
 
-        Iterable<Achievements> ach = achRepository.findAllById(achIDs);
+        return "redirect:/scientific-and-educational-centers";
+    }
 
-        if (sec.isPresent()) {
-            model.addAttribute("sec", sec.get());
-            model.addAttribute("title", sec.get().getTitle());
+    @PostMapping("/achievements/delete")
+    public String deleteAchievement(@RequestParam("id") Long id, Model model){
+
+        if (user.getAccess_level() != null){  // проверка роли
+            if(achRepository.existsById(id)) {
+                achRepository.deleteById(id);
+            }
         }
-        model.addAttribute("ach", ach);
 
-        if(secRepository.existsById(id)) {
-            SEC sec = new SEC(id, title, full_name, description);
-            secRepository.save(sec);
-
-        }*/
         return "redirect:/scientific-and-educational-centers";
     }
 
@@ -145,20 +195,27 @@ public class SECController {
                            @RequestParam("full_name") String full_name,
                            @RequestParam("description") String description,
                            Model model){
-        if(secRepository.existsById(id)) {
-            SEC sec = new SEC(id, title, full_name, description);
-            secRepository.save(sec);
 
+        if (user.getAccess_level() != null){  // проверка роли
+            if(secRepository.existsById(id)) {
+                SEC sec = new SEC(id, title, full_name, description);
+                secRepository.save(sec);
+            }
         }
+
         return "redirect:/scientific-and-educational-centers";
     }
 
     @PostMapping("/delete")
     public String deleteSEC(@RequestParam("id") Long id,
                              Model model){
-        if(secRepository.existsById(id)) {
-            secRepository.deleteById(id);
+
+        if (user.getAccess_level() != null){  // проверка роли
+            if(secRepository.existsById(id)) {
+                secRepository.deleteById(id);
+            }
         }
+
         return "redirect:/scientific-and-educational-centers";
     }
 }
